+++ src/server/game/Handlers/MiscHandler.cpp
@@ -691,7 +691,7 @@ void WorldSession::HandleAddIgnoreOpcodeCallBack(PreparedQueryResult result)
 
     if (result)
     {
-        IgnoreGuid = ObjectGuid(HIGHGUID_PLAYER, (*result)[0].GetUInt32());
+        IgnoreGuid = ObjectGuid(HIGHGUID_PLAYER, (*result)[0].GetUInt64());
 
         if (!IgnoreGuid.IsEmpty())
         {
